Creating fakers for usermeeting
UserMeeting works correctly, views needs to be created

+++ PetShop/database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,11 +19,10 @@ class CreateUsersTable extends Migration
             $table->string('address');
             $table->string('password');
             $table->string('email')->unique();
-            $table->integer('contact');
-            $table->integer('creditCard');
-            $table->integer('amount');
+            $table->bigInteger('contact');
+            $table->bigInteger('creditCard');
+            $table->bigInteger('amount');
             $table->string('petlist');
-            //$table->timestamp('email_verified_at')->nullable();
             $table->rememberToken();
             $table->timestamps();
         });
